[Backend] Promote PassPipeline to backend sub-folder and fix metal transform position
