+++ tilelang/backend/__init__.py
@@ -1,6 +1,2 @@
-# Import built-in backend packages so their pipelines register.
-from . import cpu as _cpu  # noqa: F401,E402
-from . import common as _common  # noqa: F401,E402
-from . import cuda as _cuda  # noqa: F401,E402
-from . import metal as _metal  # noqa: F401,E402
-from . import rocm as _rocm  # noqa: F401,E402
+from .pass_pipeline import PassPipeline, register_pipeline, resolve_pipeline  # noqa: F401
+from . import common as common  # noqa: F401,E402
