+++ tilelang/backend/common.py
@@ -1,7 +1,7 @@
 from __future__ import annotations
 
-from tilelang.backend.pipeline import Pipeline, register_pipeline
-from tilelang.backend.cpu.pipeline import CPUPassPipelineBody
+from tilelang.backend.pass_pipeline.pipeline import PassPipeline, register_pipeline
+from tilelang.cpu.pipeline import CPUPassPipelineBody
 
 
-register_pipeline(Pipeline("webgpu", CPUPassPipelineBody))
+register_pipeline(PassPipeline("webgpu", CPUPassPipelineBody))
