--- tilelang/backend/cpu/__init__.py
@@ -1,1 +1,0 @@
-from . import pipeline  # noqa: F401
