--- tilelang/backend/metal/__init__.py
@@ -1,2 +1,0 @@
-from . import pipeline  # noqa: F401
-from . import gemm  # noqa: F401
