+++ tilelang/backend/pass_pipeline/__init__.py
@@ -1,0 +1,8 @@
+from .pipeline import PassPipeline, get_pipeline, register_pipeline, resolve_pipeline
+
+__all__ = [
+    "PassPipeline",
+    "get_pipeline",
+    "register_pipeline",
+    "resolve_pipeline",
+]
