+++ tilelang/cpu/__init__.py
@@ -1,1 +1,2 @@
+from . import pipeline  # noqa: F401
 from . import op  # noqa: F401
