+++ tilelang/cuda/pipeline.py
@@ -1,13 +1,24 @@
 from __future__ import annotations
-from tvm import tirx, s_tir, IRModule
+
+from tvm import IRModule, s_tir, tirx
 from tvm.target import Target
+
 import tilelang
+from tilelang.backend.pass_pipeline.pipeline import PassPipeline, register_pipeline
+from tilelang.backend.pass_pipeline.pipeline_utils import (
+    LayoutVisual,
+    allow_vectorize,
+    should_disable_shared_memory_reuse,
+    should_enable_aggressive_merge,
+    should_enable_race_check,
+    should_force_let_inline,
+)
+from tilelang.contrib.nvcc import have_pdl, have_tma
 from tilelang.transform import PassContext
-from tilelang.contrib.nvcc import have_tma, have_pdl
 
 
 def allow_warp_specialized(pass_ctx: PassContext | None = None, target: Target | None = None) -> bool:
-    # avoid circular import
+    # Avoid importing jit.adapter.utils at module import time.
     from tilelang.jit.adapter.utils import is_cuda_target
 
     if pass_ctx is None:
@@ -28,142 +39,8 @@ def module_has_tma(mod: IRModule) -> bool:
     return any(func.attrs and func.attrs.get("tl.has_tma", False) for _, func in mod.functions.items())
 
 
-def allow_vectorize(pass_ctx: PassContext | None = None) -> bool:
-    if pass_ctx is None:
-        pass_ctx = tilelang.transform.get_pass_context()
-    disable_vectorize = pass_ctx.config.get("tirx.disable_vectorize", False)
-    return not disable_vectorize
-
-
-def should_enable_aggressive_merge(pass_ctx: PassContext | None = None, target: Target | None = None) -> bool:
-    if pass_ctx is None:
-        pass_ctx = tilelang.transform.get_pass_context()
-    enable_aggressive_merge = bool(pass_ctx.config.get(tilelang.PassConfigKey.TL_ENABLE_AGGRESSIVE_SHARED_MEMORY_MERGE, False))
-    if allow_warp_specialized(pass_ctx=pass_ctx, target=target):
-        # This is a workaround to avoid the bug in the MergeSharedMemoryAllocations pass
-        # when warp specialization is enabled, as different warp threads may access different
-        # buffers, but the liveness analysis is hard because we need to do pipeline.
-        enable_aggressive_merge = False
-    return enable_aggressive_merge
-
-
-def should_force_let_inline(pass_ctx: PassContext | None = None) -> bool:
-    if pass_ctx is None:
-        pass_ctx = tilelang.transform.get_pass_context()
-    return bool(pass_ctx and pass_ctx.config.get(tilelang.PassConfigKey.TL_FORCE_LET_INLINE, False))
-
-
-def should_enable_ast_print(pass_ctx: PassContext | None = None) -> bool:
-    if pass_ctx is None:
-        pass_ctx = tilelang.transform.get_pass_context()
-    return bool(pass_ctx and pass_ctx.config.get(tilelang.PassConfigKey.TL_AST_PRINT_ENABLE, False))
-
-
-def should_enable_layout_visual(pass_ctx: PassContext | None = None) -> bool:
-    if pass_ctx is None:
-        pass_ctx = tilelang.transform.get_pass_context()
-    enabled = pass_ctx.config.get(tilelang.PassConfigKey.TL_LAYOUT_VISUALIZATION_ENABLE, False)
-    return enabled
-
-
-def should_enable_race_check(pass_ctx: PassContext | None = None) -> bool:
-    if pass_ctx is None:
-        pass_ctx = tilelang.transform.get_pass_context()
-    enabled = not pass_ctx.config.get(tilelang.PassConfigKey.TL_DISABLE_DATA_RACE_CHECK, False)
-    return enabled
-
-
-def should_enable_prelower_semantic_check(pass_ctx: PassContext | None = None) -> bool:
-    if pass_ctx is None:
-        pass_ctx = tilelang.transform.get_pass_context()
-    enabled = not pass_ctx.config.get(tilelang.PassConfigKey.TL_DISABLE_PRELOWER_SEMANTIC_CHECK, False)
-    return enabled
-
-
-def should_disable_shared_memory_reuse(pass_ctx: PassContext | None = None) -> bool:
-    if pass_ctx is None:
-        pass_ctx = tilelang.transform.get_pass_context()
-    return bool(pass_ctx.config.get(tilelang.PassConfigKey.TL_DISABLE_SHARED_MEMORY_REUSE, False))
-
-
-def get_layout_visual_formats(pass_ctx: PassContext | None = None) -> list[str]:
-    if pass_ctx is None:
-        pass_ctx = tilelang.transform.get_pass_context()
-    formats_value = pass_ctx.config.get(tilelang.PassConfigKey.TL_LAYOUT_VISUALIZATION_FORMATS, "")
-    if not formats_value:
-        return ["txt"]
-
-    formats_str = formats_value.strip().lower()
-    valid_formats = ["txt", "png", "pdf", "svg", "all"]
-
-    if formats_str == "all":
-        return ["txt", "png", "pdf", "svg"]
-
-    if "," in formats_str:
-        formats_list = [f.strip() for f in formats_str.split(",")]
-    else:
-        formats_list = [formats_str]
-
-    invalid_formats = [f for f in formats_list if f not in valid_formats]
-    if invalid_formats:
-        raise ValueError(
-            f"Invalid formats for TL_LAYOUT_VISUALIZATION_FORMATS: {invalid_formats}. "
-            f"Valid formats are: {valid_formats}. "
-            f"You can choose one of the valid formats or a comma-separated list of formats.(e.g., 'txt,png,pdf')"
-        )
-    return formats_list
-
-
-def LayoutVisual(mod: IRModule) -> None:
-    """Apply layout visualization pass if enabled."""
-    if should_enable_layout_visual():
-        formats = get_layout_visual_formats()
-        tilelang.analysis.LayoutVisual(formats=formats)(mod)
-
-
-def PreLowerSemanticCheck(mod: IRModule) -> None:
-    """
-    Check whether the module is valid before lowering. If not, raise a user-friendly error
-    in Python side instead of letting the error dive into the complicated TVM/C++ stack.
-    Note: This is a validation-only pipeline of passes and does not modify or return the module.
-    """
-
-    if not should_enable_prelower_semantic_check():
-        return
-
-    # Print AST for debugging purpose
-    if should_enable_ast_print():
-        tilelang.analysis.ASTPrinter()(mod)
-    # Check if there are any invalid nested loops.
-    tilelang.analysis.NestedLoopChecker()(mod)
-    # Check if there are any invalid symbolic T.Parallel + fragment access.
-    tilelang.analysis.FragmentLoopChecker()(mod)
-
-
-def LowerAndLegalize(mod: IRModule, target: Target) -> IRModule:
-    # Bind the target device information to the module
-    """
-    Bind target information and progressively legalize and lower frontend Tile IR into a form suitable for downstream optimization and codegen.
-
-    This pass pipeline:
-    - Binds the provided target to the module.
-    - Legalizes frontend Tile IR into TVM-compatible constructs.
-    - Simplifies expressions.
-    - Configures reducer layouts and performs layout inference for fragments and shared memory.
-    - Lowers high-level tile operations and L2 persistent maps.
-    - Legalizes vectorized loops and inserts safety checks for memory accesses.
-    - Re-simplifies to remove redundancies introduced by safety checks.
-    - Attempts loop vectorization for dynamic-shaped loops.
-
-    Parameters:
-        mod (IRModule): The input IR module containing frontend Tile IR.
-        target (Target): Target device information to bind into the module.
-
-    Returns:
-        IRModule: The transformed module, ready for target-specific optimization passes.
-    """
+def CUDAPassPipelineBodyPrologue(mod: IRModule, target: Target) -> IRModule:
     mod = tirx.transform.BindTarget(target)(mod)
-
     if should_force_let_inline():
         # Force-let inline whenever the pass config requests it.
         mod = tilelang.transform.LetInline()(mod)
@@ -180,6 +57,8 @@ def LowerAndLegalize(mod: IRModule, target: Target) -> IRModule:
     mod = tilelang.transform.Simplify()(mod)
     # Set layouts for reducers
     mod = tilelang.transform.LayoutReducer()(mod)
+
+    # @CUDA-specific
     # Tile-level warp specialization: runs before layout inference so that
     # producer/consumer split happens at the high-level tile-op IR.
     # The pass classifies copy ops as TMA/cp.async/sync inline (no prior
@@ -188,28 +67,36 @@ def LowerAndLegalize(mod: IRModule, target: Target) -> IRModule:
     # applies.
     if allow_warp_specialized(target=target):
         mod = tilelang.transform.ProducerConsumerWarpSpecialized()(mod)
+
+    # @CUDA / Blackwell specific
     # Lower 2SM TCGEN5MMA and related on Blackwell target (must run before
     # LayoutInference so that the use_2cta annotation is visible to infer_layout)
     mod = tilelang.transform.LowerBlackwell2SM()(mod)
+
     # Normalize if-without-else wrappers before pipeline planning. This keeps
     # pipeline body extraction focused on canonical SeqStmt bodies.
     mod = tilelang.transform.IfStmtBinding()(mod)
+
     # Run pipeline planning and software-pipeline rewriting before layout
     # inference so inferred layouts see the final pipelined structure directly.
     mod = tilelang.transform.PipelinePlanning()(mod)
     mod = tilelang.transform.InjectSoftwarePipeline()(mod)
     mod = tilelang.transform.Simplify()(mod)
+
+    # @Metal specific
     # On Metal, rewrite local.fragment GEMM accumulators to metal.simdgroup
     # before layout inference. simdgroup matrices are opaque and have no
     # explicit thread-level layout, so layout inference must not see them.
+    # mod = tilelang.metal.transform.MetalFragmentToSimdgroup(mod)
 
-    mod = tilelang.transform.metal.MetalFragmentToSimdgroup(mod)
     # Infer memory layouts for fragments and shared memory
     mod = tilelang.transform.LayoutInference()(mod)
     # Visualize the layout
     LayoutVisual(mod)
     # Lower high-level tile operations to low-level operations
     mod = tilelang.transform.LowerTileOp()(mod)
+
+    # @CUDA specific
     # Lower l2 persistent map
     mod = tilelang.cuda.transform.LowerL2Persistent()(mod)
     # Decouple type cast vectorization constraints before vectorization
@@ -231,18 +118,27 @@ def LowerAndLegalize(mod: IRModule, target: Target) -> IRModule:
     return mod
 
 
-def OptimizeForTarget(mod: IRModule, target: Target) -> IRModule:
+def CUDAPassPipelineBody(mod: IRModule, target: Target) -> IRModule:
     pass_ctx = tilelang.transform.get_pass_context()
+
+    mod = CUDAPassPipelineBodyPrologue(mod, target)
+
+    # @CUDA-specific
     # Lower the shared.tmem into specific initialization slot
     mod = tilelang.transform.LowerSharedTmem()(mod)
-    has_tma = module_has_tma(mod)
+
     # Pipeline barriers are now created at final expanded size by
     # InjectSoftwarePipeline, so no late MVB barrier fixup is needed.
     # Buffer allocation placement is handled uniformly for both paths.
     mod = tilelang.transform.PlanAndUpdateBufferAllocationLocation()(mod)
+    # @CUDA-specific
     mod = tilelang.transform.LowerSharedBarrier()(mod)
+
+    # @CUDA-specific
+    has_tma = module_has_tma(mod)
     if has_tma:
         mod = tilelang.transform.FuseMBarrierArriveExpectTx()(mod)
+
     mod = tilelang.transform.HoistGlobalBufferAllocations()(mod)
     mod = tilelang.transform.LowerOpaqueBlock()(mod)
     mod = tilelang.transform.Simplify()(mod)
@@ -274,37 +170,58 @@ def OptimizeForTarget(mod: IRModule, target: Target) -> IRModule:
     # the Legalization.
     mod = s_tir.transform.InferFragment()(mod)
     mod = tilelang.transform.LowerThreadAllreduce()(mod)
+
+    # @CUDA-specific
     mod = tilelang.transform.LowerLDGSTG()(mod)
     mod = tilelang.cuda.transform.LowerHopperIntrin()(mod)
+
     mod = tilelang.transform.AnnotateDeviceRegions()(mod)
     mod = tilelang.transform.SplitHostDevice()(mod)
+
+    # @CUDA-specific
     # Mark the function contains pdl_sync or pdl_trigger
     mod = tilelang.transform.MarkCudaSyncCalls(have_pdl(target))(mod)
     mod = tilelang.transform.AnnotateReadOnlyParams()(mod)
+
     # MergeSharedMemoryAllocations must be applied after SplitHostDevice
     # because the merged allocation site is at the beginning of each device function
     enable_aggressive_merge = should_enable_aggressive_merge(pass_ctx=pass_ctx, target=target)
     disable_reuse = should_disable_shared_memory_reuse(pass_ctx=pass_ctx)
     mod = tilelang.transform.MergeSharedMemoryAllocations(enable_aggressive_merge=enable_aggressive_merge, disable_reuse=disable_reuse)(mod)
+
+    # @CUDA-specific
     # InjectFenceProxy is a no-op on targets that lack the TMA / async-proxy
     # programming model; the pass itself checks the PrimFunc's target.
     mod = tilelang.transform.InjectFenceProxy()(mod)
+
     mod = tilelang.transform.ThreadSync("shared")(mod)
     mod = tilelang.transform.ThreadSync("shared.dyn")(mod)
+
+    # @CUDA-specific
     # Inject conservative tcgen05 fences on Blackwell (SM100+).
     # Must run after ThreadSync so that tvm_storage_sync calls are present.
     # The pass handles shared syncs and simple linear wait/use, use/arrive
     # handoffs, and is a no-op on non-SM100 targets or functions without TMEM.
     mod = tilelang.transform.InjectTcgen05Fence()(mod)
+
     mod = tilelang.transform.MergeIfStmt()(mod)
+
+    # @CUDA-specific
     # NOTE: LowerPTXAsyncCopy is applied earlier (before PipelinePlanning).
     if allow_warp_specialized(pass_ctx=pass_ctx, target=target):
         mod = tilelang.transform.AnnotateWarpGroupRegAlloc()(mod)
+
     mod = tilelang.transform.MakePackedAPI()(mod)
     mod = tilelang.transform.Simplify()(mod)
     mod = tilelang.transform.LowerDeviceKernelLaunch()(mod)
 
+    # @CUDA-specific
     # Transform threadblock to persistent threadblock
     mod = tilelang.cuda.transform.PersistThreadblock()(mod)
 
     return mod
+
+
+cuda_pipeline = PassPipeline("cuda", CUDAPassPipelineBody)
+
+register_pipeline(cuda_pipeline)
