+++ tilelang/metal/__init__.py
@@ -1,4 +1,3 @@
-from . import intrinsics  # noqa: F401
-from . import op  # noqa: F401
+from . import gemm  # noqa: F401
 from . import pipeline  # noqa: F401
 from . import transform  # noqa: F401
