+++ tilelang/rocm/__init__.py
@@ -1,2 +1,3 @@
 from . import intrinsics  # noqa: F401
 from . import op  # noqa: F401
+from . import pipeline  # noqa: F401
